requirements.txt
web3==5.31.4
requests
python-telegram-bot==13.15